Minor Updates
- DataCore.Core.cs Load() method now accepts full filepath to data.000 or directory (relative path) that data.000 exists in
- DataCore.Core.cs Load() no longer assigns DataID for performance gain
- DataCore.Core.cs Save() useless comments removed, rewritten for IO performance benefits/readability
- DataCore.Core.cs GetEntriesByExtensions() method added to retrieve a list of entries by more than one extension
- DataCore.Functions.StringCipher.cs GetID() code rewritten for clarity/debugging purposes and to ensure accuracy
- DataCore.Functions.XOR.cs Cipher() redundant/useless (int) cast removed

+++ DataCore/Functions/XOR.cs
@@ -542,7 +542,7 @@ namespace DataCore.Functions
         {
             for (int i = 0; i < buffer.Length; i++)
             {
-                buffer[i] ^= s_CipherTable[(int)index];
+                buffer[i] ^= s_CipherTable[index];
                 index++;
             }
         }
